# openapi/amcrest-viewer.openapi.yaml
---
openapi: 3.0.2
info:
  title: Amcrest Viewer API
  version: '1.0'
servers:
- url: http://0.0.0.0:9000
paths:
  "/api/cameras":
    get:
      summary: Get all cameras
      description: Get all cameras
      operationId: getCameras
      parameters:
        - in: query
          name: latest_snapshot
          description: 'set true to return latest snapshot info'
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Camera"

  "/api/cameras/{id}":
      get:
        summary: Get all cameras
        description: Get all cameras
        operationId: getCamera
        parameters:
          - in: path
            name: id
            required: true
            description: 'camera ID'
            schema:
              type: integer
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                    "$ref": "#/components/schemas/Camera"
  "/api/cameras/{id}/live":
      get:
        summary: Get Live Stream
        description: Get camera live stream
        operationId: getCameraLiveStream
        parameters:
          - in: path
            name: id
            required: true
            description: 'camera ID'
            schema:
              type: integer
          - in: query
            name: redirect
            description: 'redirect request'
            schema:
              type: boolean
              default: true
        responses:
            '200':
              description: OK
              content:
                application/json:
                  schema:
                    type: object
                    properties:
                      uri:
                        type: string
                    required:
                      - uri
  
      
  "/api/cameras/{id}/files":
      get:
        summary: Get files
        description: Get camera files
        operationId: getCameraFiles
        parameters:
          - in: path
            name: id
            required: true
            description: 'camera ID'
            schema:
              type: string
          - in: query
            name: start
            description: 'range start'
            schema:
              type: string
              format: date-time
          - in: query
            name: end
            description: 'range end'
            schema:
              type: string
              format: date-time
          - in: query
            name: sort
            description: "sort order"
            example: desc
            schema:
              type: string
              enum:
                - asc
                - desc

        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  type: array
                  items:
                    "$ref": "#/components/schemas/CameraFile"
  "/api/cameras/{id}/stats":
    get:
      description: 'Gets camera stats for time range'
      operationId: getCameraStats
      parameters: 
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: start
        in: query
        schema:
          type: string
          format: date-time
      - name: end
        in: query
        schema:
          type: string
          format: date-time
      - name: breakdown
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
              application/json:
                schema:
                    "$ref": "#/components/schemas/CameraStats"



components:
  schemas:
    Camera:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        host:
          type: string
        last_seen:
          type: string
          format: date-time
        enabled:
          type: boolean
        timezone:
          type: string
        max_file_age_days:
          type: integer
        max_file_total_mb:
          type: integer
        username:
          type: string
        latest_snapshot:
          "$ref": "#/components/schemas/LatestSnapshot"
      required:
        - id
        - name
        - type
        - host
    LatestSnapshot:
      type: object
      properties:
        id:
          type: integer
        camera_id:
          type: integer
        path:
          type: string
        type:
          type: integer
        timestamp:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        length:
          type: integer
      required:
        - id
        - camera_id
        - path
        - type
        - timestamp
        - duration_seconds
        - length
      additionalProperties: false
    CameraFile:
      type: object
      properties:
        id:
          type: integer
        camera_id:
          type: integer
        path:
          type: string
        type:
          type: integer
        timestamp:
          type: string
          format: date-time
        duration_seconds:
          type: integer
        length:
          type: integer
      required:
        - id
        - camera_id
        - path
        - type
        - timestamp
        - duration_seconds
        - length
    CameraStats:
      type: object
      properties:
        file_count:
          type: integer
        file_size:
          type: integer
        min_date:
          type: string
          format: date-time
        max_date:
          type: string
          format: date-time
        file_counts:
          type: object
          additionalProperties:         
            "$ref": "#/components/schemas/FileData"
        breakdown:
          type: array
          items:
           "$ref": "#/components/schemas/FileData"
      required:
        - file_count
        - file_size
        - min_date
        - max_date
    FileData:
      type: object
      properties:
        count:
          type: integer
        size:
          type: integer
      required:
        - count
        - size